responsegeneration/pyproject.toml
[tool.poetry]
name = "qachatbot"
version = "0.1.0"
description = "Question-Answering Chatbot"
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
ignite = "^1.1.0"
ijson = "^3.1.4"
matplotlib = "^3.5.1"
numpy = { version = "^1.21.5", python = ">=3.8,<3.11" }
pandas = "^1.3.5"
pytorch-ignite = "^0.4.7"
torch = "^1.10.1"
tqdm = "^4.62.3"
transformers = "^4.7.0"
tensorboardX = "^2.4.1"
farm-haystack = "^1.0.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
